List Clones

+++ CloneWar/src/main/java/fr/uge/clonewar/backend/model/Clones.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Objects;
 
 public record Clones(Artefact reference, List<Clone> clones) {
-  public record Clone(Artefact artefact, double percentage) {
+  public record Clone(Artefact artefact, int percentage) {
     public Clone {
       Objects.requireNonNull(artefact);
     }
